+++ CHANGES.md
@@ -209,3 +209,4 @@
 - Bernafon/Sonic/Philips: target market accepts manually input `default` rather than only accepts pressing Enter with default
 - ReSound/Beltone/Interton: add an option to manually specify a version
 - Clean screen before menus
+- Layout adjustment