+++ CHANGES.md
@@ -205,7 +205,7 @@
 - Unitron: fix TrueFit checker for v5.7.1 by changing default target market
 - Hansaton: fix scout checker for v5.7.1 by changing default target market
 
-### Pre-release (work-in-progress)
+### v2025.04.21
 - Bernafon/Sonic/Philips: target market accepts manually input `default` rather than only accepts pressing Enter with default
 - ReSound/Beltone/Interton: add an option to manually specify a version
 - Widex: option to change target market (country)